% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{bz.install}
\alias{bz.install}
\title{Install a package from bioconductor}
\usage{
bz.install(pkg)
}
\arguments{
\item{pkg}{eg, 'biomaRt'}
}
\value{
returns nothing
}
\description{
internally source/install biocLite.R, wrapper of biocLite(pkg)
}

