% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.geneinfo}
\alias{mart.geneinfo}
\title{retrieve human gene info from ensembl}
\usage{
mart.geneinfo(values, filters = "ensembl_gene_id",
  attributes = c("ensembl_gene_id", "hgnc_symbol", "description"),
  host = NULL)
}
\arguments{
\item{values}{the actual input data values, ensembl_gene_id, ENSG00000118473, c('ENSG00000118473', 'ENSG00000162426')}

\item{filters}{the kind/type of your input data}

\item{attributes}{what to return, max=3(?), otherwise, Too many attributes selected for External References
\cr "hgnc_id","entrezgene","kegg_enzyme","go_id","ucsc"}

\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'May2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\value{
returns a data frame
}
\description{
retrieve human gene info from ensembl
}
\note{
In case multple filters are in use, the values argument requires a list of values where each position in the list corresponds to the position of the filters in the filters argument. eg,
\cr filters=c("chr_name","start","end")
\cr values=list(8,148350, 158612)
}

