% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.genesnps}
\alias{mart.genesnps}
\title{retrieves a gene's all snps}
\usage{
mart.genesnps(values = "CRHR1", filters = "hgnc_symbol",
  attributes = c("variation_name", "validated", "allele", "minor_allele",
  "minor_allele_freq", "transcript_count", "clinical_significance",
  "ensembl_gene_id", "chromosome_name", "start_position", "end_position",
  "band", "external_gene_name", "description"), host = NULL)
}
\arguments{
\item{values}{the actual input data values, ensembl_gene_id, ENSG00000118473, c('ENSG00000118473', 'ENSG00000162426'). If vector, should be the same id type}

\item{filters}{input data type, 'hgnc_symbol'
\cr "ensembl_gene_id"(ENSG00000118473),"hgnc_id"(HGNC:25412),"entrezgene" (84251),"kegg_enzyme"(00010+1.1.1.1),"go_id"(GO:0030122),"ucsc"(uc057hhx.1)}

\item{attributes}{what to return}

\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'May2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\value{
returns a data frame
}
\description{
retrieves a gene's all snps
}

