% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.list}
\alias{mart.list}
\title{Lists available archived versions of Ensembl (hosts), marts, snp attributes/filters, gene attributes/filters}
\usage{
mart.list(what = "snp", host = NULL, ...)
}
\arguments{
\item{what}{one of 'hosts', 'marts', 'snp', 'gene'}

\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'May2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\value{
view in RStudio
}
\description{
Lists available archived versions of Ensembl (hosts), marts, snp attributes/filters, gene attributes/filters
}

