% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.snp}
\alias{mart.snp}
\title{return a mart object representing snp:hsapiens_snp}
\usage{
mart.snp(host = NULL, biomart = "snp", dataset = "hsapiens_snp", ...)
}
\arguments{
\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'aug2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\description{
return a mart object representing snp:hsapiens_snp
}
