% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.snpinfo}
\alias{mart.snpinfo}
\title{retrieve human snp info from ensembl}
\usage{
mart.snpinfo(values, filters = "snp_filter", attributes = c("refsnp_id",
  "chr_name", "chrom_start", "chrom_end", "allele", "allele_1", "minor_allele",
  "minor_allele_freq", "synonym_name"), host = NULL)
}
\arguments{
\item{values}{rs number, eg, 'rs2075507'}

\item{filters}{'snp_filter'}

\item{attributes}{c('refsnp_id','chr_name','chrom_start','chrom_end','allele',
'allele_1','minor_allele','minor_allele_freq','synonym_name')}

\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'May2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\value{
returns a data frame
}
\description{
retrieve human snp info from ensembl
}

