% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{snp.isnotbiallelic}
\alias{snp.isnotbiallelic}
\title{the cols that are not biallelic}
\usage{
snp.isnotbiallelic(df, cols = NULL)
}
\arguments{
\item{df}{df}

\item{cols}{evaluated as cols=ez.selcol(x,cols). NULL=all cols}
}
\value{
returns a vector of names of column that is not biallelic
}
\description{
the cols that are not biallelic
}
