% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp.R
\name{snp.recode}
\alias{snp.recode}
\title{recode snps}
\usage{
snp.recode(x, cols = NULL, recodes = c(0, 1, 2))
}
\arguments{
\item{x}{df or vector (char or factor of char)}

\item{cols}{evaluated as cols=ez.selcol(x,cols), ignored if x is a vector}

\item{recodes}{vector with three elements corresponding to (minor-minor, minor-major or major-minor, major-major). Could be c(0,1,2), c(-1,0,1), c(1,2,3), c(0,1,1), c(1,0,0), c('AA','AB','BB')}
}
\value{
returns a new df or vector. Regardless of input data type, if recodes are number, then the returned is also numeric; if recodes are char, then always factor of char.
}
\description{
recode snps
}
\note{
assume biallelic, otherwise error. minor/major bases are calculated based on the actual input data
}
