% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{bz.install}
\alias{bz.install}
\title{Install a package from bioconductor}
\usage{
bz.install(pkg, suppressUpdates = TRUE, ...)
}
\arguments{
\item{pkg}{eg, 'biomaRt'}

\item{suppressUpdates}{if TRUE, do not ask to update other old packages and do not update them; if FALSE, will ask to choose}
}
\value{
returns nothing
}
\description{
internally source/install biocLite.R, wrapper of \code{\link[BiocInstaller]{biocLite}}
}
