% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.geneinfo}
\alias{mart.geneinfo}
\title{retrieve human gene info from ensembl}
\usage{
mart.geneinfo(values = c("ENSG00000118473", "ENSG00000162426"),
  filters = "ensembl_gene_id", attributes = c("ensembl_gene_id",
  "hgnc_symbol", "description", "chromosome_name", "start_position",
  "end_position"), host = NULL)
}
\arguments{
\item{values}{the actual input data values, ensembl_gene_id, ENSG00000118473, c('ENSG00000118473', 'ENSG00000162426'). If vector, should be the same id type}

\item{filters}{the kind/type of your input data, 'hgnc_symbol'
\cr "ensembl_gene_id"(ENSG00000118473),"hgnc_id"(HGNC:25412),"entrezgene" (84251),"kegg_enzyme"(00010+1.1.1.1),"go_id"(GO:0030122),"ucsc"(uc057hhx.1)}

\item{attributes}{what to return}

\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'May2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\value{
returns a data frame
}
\description{
retrieve human gene info from ensembl
}
\note{
In case multple filters are in use, the values argument requires a list of values where each position in the list corresponds to the position of the filters in the filters argument. eg,
\cr filters=c("chr_name","start","end")
\cr values=list(8,148350, 158612)
}

