% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{mart.snpinfo}
\alias{mart.snpinfo}
\alias{mart.snpinfo2}
\title{retrieve human snp info from ensembl}
\usage{
mart.snpinfo(values, filters = "snp_filter", attributes = c("refsnp_id",
  "chr_name", "chrom_start", "chrom_end", "allele", "allele_1", "minor_allele",
  "minor_allele_freq", "synonym_name", "ensembl_gene_stable_id"), host = NULL)

mart.snpinfo2(values, filters = "snp_synonym_filter",
  attributes = c("refsnp_id", "chr_name", "chrom_start", "chrom_end",
  "allele", "allele_1", "minor_allele", "minor_allele_freq", "synonym_name",
  "ensembl_gene_stable_id"), host = NULL)
}
\arguments{
\item{values}{the actual input data values, rs number, eg, 'rs2075507'; to search archived synonymous rs number, use mart.snpinfo2}

\item{filters}{the kind/type of your input data, eg, 'snp_filter'}

\item{attributes}{what to return}

\item{host}{default 'www.ensembl.org'. Other eg, 'grch37.ensembl.org', 'May2017.archive.ensembl.org'. See all, run \code{\link{mart.list}}}
}
\value{
returns a data frame
}
\description{
retrieve human snp info from ensembl
}
\note{
In case multple filters are in use, the values argument requires a list of values where each position in the list corresponds to the position of the filters in the filters argument
filters=c("chr_name","start","end"),
values=list(8,148350, 158612),
}

